updated console logs to error where appropriate for messages. Ammended spy log tests

+++ src/api/controllers/users.ts
@@ -3,6 +3,7 @@ import userRepository from '../repositories/user/UserRepositoryFactory';
 import { IUserRepository } from '../repositories/user/IUserRepository';
 import { UserFilterParams } from '../types';
 import { stringToBoolean } from '../helpers/booleanConvert';
+import { isValidString } from '../helpers/checker';
 
 let userRepo: IUserRepository;
 
@@ -20,7 +21,7 @@ export const getAllUsers = async (req: Request, res: Response, next: NextFunctio
     
     res.status(200).json(users).end();
   } catch(error) {
-    console.log(error);
+    console.error(error);
     next(error);
   }
 }
@@ -29,13 +30,24 @@ export const deleteUser = async (req: Request, res: Response, next: NextFunction
   try {
     const { id } = req.params;
 
+    if (!isValidString(id)) {
+      res.status(400).json({ message: `Invalid id: "${id}"` })
+      return
+    }
+
+    const user = await userRepo.getUserById(id);
+    if(!user) {
+      res.status(404).send(`User with id: ${id} not found`);
+      return;
+    }
+
     await userRepo.deleteUserById(id);
     
     console.log(`Deleted user with id: ${id}`);
     res.status(204).end();
     return;
   } catch(error) {
-    console.log(error);
+    console.error(error);
     next(error);
   }
 }
@@ -43,7 +55,12 @@ export const deleteUser = async (req: Request, res: Response, next: NextFunction
 export const getUser = async (req: Request, res: Response, next: NextFunction) : Promise<void> => {
   try {
     const { id } = req.params;
-    let user = await userRepo.getUserById(id);
+
+    if (!isValidString(id)) {
+      res.status(400).json({ message: `Invalid id: "${id}"` })
+      return
+    }
+    const user = await userRepo.getUserById(id);
 
     if(!user) {
       res.status(404).send(`User with id: ${id} not found`);
@@ -52,7 +69,7 @@ export const getUser = async (req: Request, res: Response, next: NextFunction) :
     
     res.status(200).json(user).end();
   } catch(error) {
-    console.log(error);
+    console.error(error);
     next(error);
   }
 }
